Corrigido os nomes reservados usados em tabelas, corrigida as relações entre as tabelas e corrigido o problema em relação ao looping infinito ao gerar o JSON

+++ patrimonio-back-end/src/main/resources/db/migration/V04__create-Log-table.sql
@@ -3,5 +3,5 @@ CREATE TABLE [dbo].[Log](
 	[date] [date] NOT NULL,
 	[tablename] [varchar](25) NOT NULL,
 	[action] [varchar](10) NOT NULL,
-	[user_id] [int] NOT NULL FOREIGN KEY REFERENCES [User](user_id)
+	[user_id] [bigint] NOT NULL FOREIGN KEY REFERENCES [User](user_id)
 );
