+++ patrimonio-back-end/src/main/resources/db/migration/V05__create-Patrimony-table.sql
@@ -8,6 +8,6 @@ CREATE TABLE [dbo].[Patrimony](
 	[brand] [varchar](50) NULL,
 	[additional_information] [varchar](200) NULL,
 	[value] [decimal](10, 2) NULL,
-	[location_id] [int] NULL FOREIGN KEY REFERENCES Location(location_id),
-	[acquisition_method_id] [int] NULL FOREIGN KEY REFERENCES AcquisitionMethod(acquisition_method_id)
+	[location_id] [bigint] NULL FOREIGN KEY REFERENCES Location(location_id),
+	[acquisition_method_id] [bigint] NULL FOREIGN KEY REFERENCES AcquisitionMethod(acquisition_method_id)
 );
